+++ .buildkite/scripts/publish.sh
@@ -3,20 +3,34 @@ source .buildkite/scripts/tooling.sh
 
 set -euo pipefail
 
+build_docker_image() {
+    local go_version
+    go_version=$(cat .go-version)
 
-pushDockerImage() {
-    docker build \
+    docker buildx build "$@" \
+        --platform linux/amd64,linux/arm64/v8 \
         -t "${DOCKER_IMG_TAG}" \
+        -t "${DOCKER_IMG_TAG_BRANCH}" \
+        --build-arg GO_VERSION="${go_version}" \
+        --build-arg BUILDER_IMAGE=docker.elastic.co/wolfi/go \
+        --build-arg RUNNER_IMAGE=docker.elastic.co/wolfi/chainguard-base \
         --label BRANCH_NAME="${TAG_NAME}" \
         --label GIT_SHA="${BUILDKITE_COMMIT}" \
         --label GO_VERSION="${SETUP_GOLANG_VERSION}" \
         --label TIMESTAMP="$(date +%Y-%m-%d_%H:%M)" \
         .
-    retry 3 docker push "${DOCKER_IMG_TAG}"
-    echo "Docker image pushed: ${DOCKER_IMG_TAG}"
-    docker tag "${DOCKER_IMG_TAG}" "${DOCKER_IMG_TAG_BRANCH}"
-    retry 3 docker push "${DOCKER_IMG_TAG_BRANCH}"
-    echo "Docker image pushed: ${DOCKER_IMG_TAG_BRANCH}"
+}
+
+push_docker_image() {
+    docker buildx create --use
+
+    # first build the image without push
+    build_docker_image
+
+    # essentially the same as above with --push flag; the build should be in the cache
+    retry 3 build_docker_image --push
+
+    echo "Docker images pushed: ${DOCKER_IMG_TAG} ${DOCKER_IMG_TAG_BRANCH}"
 }
 
 if [[ "${BUILDKITE_PULL_REQUEST}" != "false" ]]; then
@@ -35,4 +49,4 @@ fi
 DOCKER_IMG_TAG="${DOCKER_NAMESPACE}:${BUILDKITE_COMMIT}"
 DOCKER_IMG_TAG_BRANCH="${DOCKER_NAMESPACE}:${TAG_NAME}"
 
-pushDockerImage
+push_docker_image