+++ artifacts.go
@@ -5,12 +5,13 @@
 package main
 
 import (
+	"errors"
 	"net/http"
 	"time"
 
 	"github.com/Masterminds/semver/v3"
 	"github.com/gorilla/mux"
-	"github.com/pkg/errors"
+
 	"go.elastic.co/apm/module/apmzap/v2"
 	"go.uber.org/zap"