+++ CHANGELOG.md
@@ -10,8 +10,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Bugfixes
 
+* Fix context propagation in APM transaction for watcher backend process. [#1150](https://github.com/elastic/package-registry/pull/1150) [#1152](https://github.com/elastic/package-registry/pull/1152)
+* Update Go Runtime to 1.22.2. [#1170](https://github.com/elastic/package-registry/issues/1170)
+
 ### Added
 
+* Add support for multi-platform container images. [#1162](https://github.com/elastic/package-registry/pull/1162)
+* Use Wolfi as base for container images. [#1169](https://github.com/elastic/package-registry/pull/1169)
 * Reuse HTTP client when proxifying resolver requests. [#1147](https://github.com/elastic/package-registry/pull/1147)
 
 ### Deprecated